lab4/user-service/requirements.txt
fastapi
uvicorn
pyjwt
python-multipart
passlib
psycopg2-binary
asyncpg
bcrypt